refactor: move base/LogLevel -> common/LogLevel
