+++ src/lib/base/CMakeLists.txt
@@ -25,7 +25,6 @@ add_library(base STATIC
   LogOutputters.h
   Log.cpp
   Log.h
-  LogLevel.h
   PriorityQueue.h
   SimpleEventQueueBuffer.cpp
   SimpleEventQueueBuffer.h
