+++ src/lib/platform/MSWindowsWatchdog.cpp
@@ -9,10 +9,10 @@
 #include "arch/Arch.h"
 #include "arch/win32/XArchWindows.h"
 #include "base/Log.h"
-#include "base/LogLevel.h"
 #include "base/LogOutputters.h"
 #include "base/TMethodJob.h"
 #include "common/Constants.h"
+#include "common/LogLevel.h"
 #include "deskflow/App.h"
 #include "mt/Thread.h"
 #include "platform/MSWindowsHandle.h"
